complete other utilities

+++ src/include/miopen/gather/problem_description.hpp
@@ -26,7 +26,9 @@
 
 #pragma once
 
+#include <cstddef>
 #include <cstdint>
+
 #include <miopen/errors.hpp>
 #include "miopen/gather.hpp"
 #include <miopen/tensor.hpp>
@@ -40,13 +42,13 @@ struct NetworkConfig;
 namespace gather {
 
 template <int M>
-tensor_view_t<M> reshape(const TensorDescriptor& tensorDes, const std::vector<int64_t>& shape);
+tensor_view_t<M> reshape(const TensorDescriptor& tensorDes, const std::vector<size_t>& shape);
 
 extern template tensor_view_t<3> reshape<3>(const TensorDescriptor& tensorDes,
-                                            const std::vector<int64_t>& shape);
+                                            const std::vector<size_t>& shape);
 
 extern template tensor_view_t<4> reshape<4>(const TensorDescriptor& tensorDes,
-                                            const std::vector<int64_t>& shape);
+                                            const std::vector<size_t>& shape);
 
 struct BwdProblemDescription : ProblemDescriptionBase
 {
